requirements.txt
matplotlib
pandas
pytorch-argus
albumentations
scikit-learn
git+https://github.com/rwightman/pytorch-image-models@master
